+++ src/worker/background.ts
@@ -1,35 +1,14 @@
-import { SUPABASE_LOGOUT_URL, SUPABASE_AUTH_DOMAIN, SUPABASE_COOKIE_NAME, OPEN_SAUCED_AUTH_TOKEN_KEY } from "../constants";
-import { checkTokenValidity } from "../utils/fetchOpenSaucedApiData";
-import setAccessTokenInChromeStorage from "../utils/setAccessToken";
+import { checkAuthentication } from "../utils/checkAuthentication";
+import { SUPABASE_AUTH_COOKIE_NAME, OPEN_SAUCED_INSIGHTS_DOMAIN } from "../constants";
 
-chrome.webRequest.onCompleted.addListener(
-  () => {
-    void chrome.storage.sync.remove(OPEN_SAUCED_AUTH_TOKEN_KEY);
-  },
-  { urls: [SUPABASE_LOGOUT_URL] },
-);
-
-chrome.tabs.onUpdated.addListener((tabId, changeInfo) => {
-  if (changeInfo.url?.includes("github.com")) {
-    void chrome.tabs.sendMessage(tabId, { message: "GITHUB_URL_CHANGED" });
-  }
+chrome.cookies.onChanged.addListener(changeInfo => {
+    if (
+      changeInfo.cookie.name === SUPABASE_AUTH_COOKIE_NAME ||
+      changeInfo.cookie.domain === OPEN_SAUCED_INSIGHTS_DOMAIN
+    ) {
+      checkAuthentication();
+    }
 });
 
-chrome.cookies.onChanged.addListener(async changeInfo => {
-  try {
-    if (changeInfo.cookie.name !== SUPABASE_COOKIE_NAME || changeInfo.cookie.domain !== SUPABASE_AUTH_DOMAIN) {
-      return;
-    }
-    if (changeInfo.removed) {
-      return chrome.storage.sync.remove(OPEN_SAUCED_AUTH_TOKEN_KEY);
-    }
-    const isValidToken = await checkTokenValidity(changeInfo.cookie.value);
-
-    if (!isValidToken) {
-      return chrome.storage.sync.remove(OPEN_SAUCED_AUTH_TOKEN_KEY);
-    }
-    void setAccessTokenInChromeStorage(changeInfo.cookie.value);
-  } catch (error) {
-    console.error("Error processing cookie update:", error);
-  }
-});
+chrome.runtime.onInstalled.addListener(checkAuthentication);
+chrome.runtime.onStartup.addListener(checkAuthentication);